Trading

+++ website/src/Controller/DefaultController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller;
 
+use App\Service\UserServices;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use App\Form\UploadAvatar;
@@ -94,6 +95,14 @@ class DefaultController extends AbstractController
     {
         $user = $this->getUser();
 
+        if(!$user) return new Response(
+            '<html><body>Error!</body></html>'
+        );
+
+        $sent = $this->getDoctrine()
+            ->getRepository(Trade::class)
+            ->findBy(array('userFrom' => $user->getId(), 'status' => 0));
+
         $active = $this->getDoctrine()
             ->getRepository(Trade::class)
             ->findBy(array('userTo' => $user->getId(), 'status' => 0));
@@ -103,6 +112,7 @@ class DefaultController extends AbstractController
             ->findBy(array('userTo' => $user->getId(), 'status' => array(1, 2)));
 
         return $this->render('pages/trading.html.twig', [
+            'sent' => $sent,
             'active' => $active,
             'completed' => $completed,
         ]);
@@ -114,6 +124,10 @@ class DefaultController extends AbstractController
     public function trade($id)
     {
         $user = $this->getUser();
+
+        if(!$user) return new Response(
+            '<html><body>Error!</body></html>'
+        );
 
         $trade = $this->getDoctrine()
             ->getRepository(Trade::class)
@@ -133,6 +147,10 @@ class DefaultController extends AbstractController
     {
         $user = $this->getUser();
 
+        if(!$user) return new Response(
+            '<html><body>Error!</body></html>'
+        );
+
         $trade = $this->getDoctrine()
             ->getRepository(Trade::class)
             ->findOneBy(array('userTo' => $user->getId(), 'id' => $id, 'status' => 0));
@@ -145,6 +163,10 @@ class DefaultController extends AbstractController
 
         /* ----- */
 
+        if(!$trade) return new Response(
+            '<html><body>This trade offer is not active!</body></html>'
+        );
+
         if ($trade->getFrom() == $user->getId()) {
             $receiving = $trade->getReceivingItems();
             $offering = $trade->getOfferedItems();
@@ -155,33 +177,51 @@ class DefaultController extends AbstractController
 
         /* ----- */
 
-        $values = "";
-        foreach($receiving as $item){
-            $values .= '("'.$item.'", "'.$user->getId().'"),';
-            $sql = "DELETE FROM inventory WHERE item=".$item." AND user=".$trade->getFrom()." LIMIT 1";
-            $em = $this->getDoctrine()->getManager();
-            $stmt = $em->getConnection()->prepare($sql);
-            $stmt->execute();
+        foreach ($receiving as $item) {
+           if(!$item) {
+               $receiving = false;
+           }
         }
-        $sql = "INSERT INTO inventory (item, user) VALUES ".rtrim($values, ',');
-        $em = $this->getDoctrine()->getManager();
-        $stmt = $em->getConnection()->prepare($sql);
-        $stmt->execute();
+
+        foreach ($offering as $item) {
+            if(!$item) {
+                $offering = false;
+            }
+        }
 
         /* ----- */
 
-        $values = "";
-        foreach($offering as $item){
-            $values .= '("'.$item.'", "'.$trade->getFrom().'"),';
-            $sql = "DELETE FROM inventory WHERE item=".$item." AND user=".$user->getId()." LIMIT 1";
+        if($receiving) {
+            $values = "";
+            foreach ($receiving as $item) {
+                $values .= '("' . $item . '", "' . $user->getId() . '"),';
+                $sql = "DELETE FROM inventory WHERE item=" . $item . " AND user=" . $trade->getFrom() . " LIMIT 1";
+                $em = $this->getDoctrine()->getManager();
+                $stmt = $em->getConnection()->prepare($sql);
+                $stmt->execute();
+            }
+            $sql = "INSERT INTO inventory (item, user) VALUES " . rtrim($values, ',');
             $em = $this->getDoctrine()->getManager();
             $stmt = $em->getConnection()->prepare($sql);
             $stmt->execute();
         }
-        $sql = "INSERT INTO inventory (item, user) VALUES ".rtrim($values, ',');
-        $em = $this->getDoctrine()->getManager();
-        $stmt = $em->getConnection()->prepare($sql);
-        $stmt->execute();
+
+        /* ----- */
+
+        if($offering) {
+            $values = "";
+            foreach ($offering as $item) {
+                $values .= '("' . $item . '", "' . $trade->getFrom() . '"),';
+                $sql = "DELETE FROM inventory WHERE item=" . $item . " AND user=" . $user->getId() . " LIMIT 1";
+                $em = $this->getDoctrine()->getManager();
+                $stmt = $em->getConnection()->prepare($sql);
+                $stmt->execute();
+            }
+            $sql = "INSERT INTO inventory (item, user) VALUES " . rtrim($values, ',');
+            $em = $this->getDoctrine()->getManager();
+            $stmt = $em->getConnection()->prepare($sql);
+            $stmt->execute();
+        }
 
         /* ----- */
 
@@ -202,19 +242,182 @@ class DefaultController extends AbstractController
     {
         $user = $this->getUser();
 
+        if(!$user) return new Response(
+            '<html><body>Error!</body></html>'
+        );
+
         $trade = $this->getDoctrine()
             ->getRepository(Trade::class)
             ->findOneBy(array('userTo' => $user->getId(), 'id' => $id, 'status' => 0));
 
-        $entityManager = $this->getDoctrine()->getManager();
-        $trade->setStatus(2); // Decline
-        $entityManager->flush();
+        if($trade) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $trade->setStatus(2); // Decline
+            $entityManager->flush();
+        }
 
         return $this->render('pages/trade_action.html.twig', [
             'title' => 'Trade Offer #'.$id,
             'trade' => $trade,
             'action' => 'decline'
         ]);
+    }
+
+    /**
+     * @Route("/trade/cancel/{id}", name="trade_cancel")
+     */
+    public function trade_cancel($id)
+    {
+        $user = $this->getUser();
+
+        if(!$user) return new Response(
+            '<html><body>Error!</body></html>'
+        );
+
+        $trade = $this->getDoctrine()
+            ->getRepository(Trade::class)
+            ->findOneBy(array('userFrom' => $user->getId(), 'id' => $id, 'status' => 0));
+
+        if($trade) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $trade->setStatus(3); // Cancel
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('hub');
+    }
+
+    /**
+     * @Route("/sendtrade/{usernameTo}", name="send_trade")
+     */
+    public function send_trade($usernameTo = "")
+    {
+        $user = $this->getUser();
+
+        if(!$user) return new Response(
+            '<html><body>Error!</body></html>'
+        );
+
+        $myInventory = $this->getDoctrine()
+            ->getRepository(Inventory::class)
+            ->findBy(array('user' => $user->getId()));
+
+        $myInventoryArray = array();
+        foreach($myInventory as $data){
+            array_push($myInventoryArray, $data->getItem());
+        }
+
+        $hisID = $this->getDoctrine()
+            ->getRepository(User::class)
+            ->findOneBy(array('username' =>  $usernameTo));
+
+        if(!$hisID) return new Response(
+            '<html><body>Username not found!</body></html>'
+        );
+
+        $hisInventory = $this->getDoctrine()
+            ->getRepository(Inventory::class)
+            ->findBy(array('user' =>  $hisID->getId()));
+
+        $hisInventoryArray = array();
+        foreach($hisInventory as $data){
+            array_push($hisInventoryArray, $data->getItem());
+        }
+
+        $items = $this->getDoctrine()
+            ->getRepository(Item::class)
+            ->findAll();
+
+        return $this->render('pages/send_trade.html.twig', [
+            'usernameTo' => $usernameTo,
+            'idTo' => $hisID->getId(),
+            'title' => 'Send trade to '.$usernameTo,
+            'myInventory' => $myInventoryArray,
+            'hisInventory' => $hisInventoryArray,
+            'items_list' => $items
+        ]);
+    }
+
+    /**
+     * @Route("/send-trade", name="send_trade_send")
+     */
+    public function send_trade_send(Request $request)
+    {
+        $user = $this->getUser();
+
+        if(!$user) return new Response(
+            '<html><body>Error!</body></html>'
+        );
+
+        $sentTradesCheck = $this->getDoctrine()
+            ->getRepository(Trade::class)
+            ->findBy(array('userFrom' =>  $user->getId(), 'status' => 0));
+
+        if($sentTradesCheck) return new Response(
+            '<html><body>You cant send more than one trade offer!</body></html>'
+        );
+
+        $userToID = $request->query->get('to');
+        $myItems = $request->query->get('my');
+        $theirItems = $request->query->get('their');
+
+        $hisInvCheck = explode(",", $theirItems);
+        $myInvCheck = explode(",", $myItems);
+
+        if(!$userToID) return new Response(
+            '<html><body>Missing some parameters!</body></html>'
+        );
+
+        if($theirItems) {
+            $hisInventory = $this->getDoctrine()
+                ->getRepository(Inventory::class)
+                ->findBy(array('user' => $userToID));
+
+            $hisInvArray = [];
+            foreach ($hisInventory as $data) {
+                array_push($hisInvArray, $data->getItem());
+            }
+
+            foreach (explode(",", $theirItems) as $data) {
+                $key = array_search($data, $hisInvArray);
+                if ($key !== false) {
+                    unset($hisInvArray[$key]);
+                } else return new Response(
+                    '<html><body>Receiver is missing items!</body></html>'
+                );
+            }
+        }
+        if($myItems) {
+            $myInventory = $this->getDoctrine()
+                ->getRepository(Inventory::class)
+                ->findBy(array('user' => $user->getId()));
+
+            $myInvArray = [];
+            foreach ($myInventory as $data) {
+                array_push($myInvArray, $data->getItem());
+            }
+
+            foreach (explode(",", $myItems) as $data) {
+                $key = array_search($data, $myInvArray);
+                if ($key !== false) {
+                    unset($myInvArray[$key]);
+                } else return new Response(
+                    '<html><body>Sender is missing items!</body></html>'
+                );
+            }
+        }
+        $trade = new Trade();
+        $trade->setUserFrom($user->getId());
+        $trade->setUserTo($userToID);
+        $trade->setOfferedItems($myInvCheck);
+        $trade->setReceivingItems($hisInvCheck);
+        $trade->setStatus(0);
+        $trade->setTimestamp(new \DateTime());
+        $em = $this->getDoctrine()->getManager();
+        $em->persist($trade);
+        $em->flush();
+
+        return $this->redirectToRoute('hub');
     }
 
     /**
